+++ Libplanet/Store/LiteDBStore.cs
@@ -28,7 +28,7 @@ namespace Libplanet.Store
 
         private const string StateRefIdPrefix = "stateref/";
 
-        private const string NonceIdPrefix = "nonce/";
+        private const string NonceIdPrefix = "nonce_";
 
         private readonly LiteDatabase _db;
 
@@ -354,54 +354,50 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public HashDigest<SHA256>? LookupStateReference<T>(
+        public IEnumerable<Tuple<HashDigest<SHA256>, long>> IterateStateReferences(
             string @namespace,
-            Address address,
-            Block<T> lookupUntil)
-            where T : IAction, new()
+            Address address)
         {
             var fileId = $"{StateRefIdPrefix}{@namespace}/{address.ToHex()}";
             LiteFileInfo file = _db.FileStorage.FindById(fileId);
 
             if (file is null || file.Length == 0)
             {
-                return null;
+                yield break;
+            }
+
+            int hashSize = HashDigest<SHA256>.Size;
+            int stateReferenceSize = hashSize + sizeof(long);
+            if (file.Length % stateReferenceSize != 0)
+            {
+                throw new FileLoadException(
+                    $"State references file's size ({file.Length}) should be multiple of " +
+                    $"state reference entry size {stateReferenceSize})."
+                );
             }
 
             using (var stream = new MemoryStream())
             {
+                // Note that a stream made by file.OpenRead() does not support
+                // .Seek() operation --- although it implements the interface,
+                // the method throws a NotSupportedException.
                 file.CopyTo(stream);
 
-                int hashSize = HashDigest<SHA256>.Size;
-                int stateReferenceSize = hashSize + sizeof(long);
                 var buffer = new byte[stateReferenceSize];
-
-                if (stream.Length % stateReferenceSize != 0)
-                {
-                    throw new FileLoadException(
-                        $"State reference file size {stream.Length} " +
-                        "should be multiple of state reference entry size " +
-                        $"{stateReferenceSize}");
-                }
-
                 long position = stream.Seek(0, SeekOrigin.End);
 
                 for (var i = 1; position - buffer.Length >= 0; i++)
                 {
-                    position = stream.Seek(
-                        -buffer.Length * i, SeekOrigin.End);
+                    position = stream.Seek(-buffer.Length * i, SeekOrigin.End);
                     stream.Read(buffer, 0, buffer.Length);
                     byte[] hashBytes = buffer.Take(hashSize).ToArray();
                     long index = BitConverter.ToInt64(buffer, hashSize);
-
-                    if (index <= lookupUntil.Index)
-                    {
-                        return new HashDigest<SHA256>(hashBytes);
-                    }
+                    yield return Tuple.Create(
+                        new HashDigest<SHA256>(hashBytes),
+                        index
+                    );
                 }
             }
-
-            return null;
         }
 
         /// <inheritdoc/>
@@ -511,146 +507,21 @@ namespace Libplanet.Store
         /// <inheritdoc/>
         public long GetTxNonce(string @namespace, Address address)
         {
-            var fileId = $"{NonceIdPrefix}{@namespace}/{address.ToHex()}";
-            LiteFileInfo file = _db.FileStorage.FindById(fileId);
-
-            if (file is null || file.Length == 0)
-            {
-                return 0;
-            }
-
-            int hashSize = HashDigest<SHA256>.Size;
-            int blockIndexSize = sizeof(long);
-            int nonceSize = sizeof(long);
-            int nonceEntrySize = hashSize + blockIndexSize + nonceSize;
-
-            using (var stream = new MemoryStream())
-            {
-                file.CopyTo(stream);
-
-                if (stream.Length % nonceEntrySize != 0)
-                {
-                    throw new FileLoadException(
-                        $"Nonce file size {stream.Length} should be " +
-                        $"a multiple of nonce entry size {nonceEntrySize}");
-                }
-
-                var buffer = new byte[nonceEntrySize];
-                stream.Seek(stream.Length - buffer.Length, SeekOrigin.Begin);
-                stream.Read(buffer, 0, buffer.Length);
-
-                return BitConverter.ToInt64(buffer, hashSize + blockIndexSize);
-            }
+            var collectionId = $"{NonceIdPrefix}{@namespace}";
+            LiteCollection<BsonDocument> collection = _db.GetCollection<BsonDocument>(collectionId);
+            var docId = new BsonValue(address.ToByteArray());
+            BsonDocument doc = collection.FindById(docId);
+            return doc is null ? 0 : (doc.TryGetValue("v", out BsonValue v) ? v.AsInt64 : 0);
         }
 
         /// <inheritdoc/>
-        public void IncreaseTxNonce<T>(string @namespace, Block<T> block)
-            where T : IAction, new()
+        public void IncreaseTxNonce(string @namespace, Address signer, long delta = 1)
         {
-            IEnumerable<Address> signers = block
-                .Transactions
-                .Select(tx => tx.Signer);
-            int hashSize = HashDigest<SHA256>.Size;
-            long blockIndex = block.Index;
-
-            foreach (Address signer in signers)
-            {
-                var fileId = $"{NonceIdPrefix}{@namespace}/{signer.ToHex()}";
-                long nextNonce = GetTxNonce(@namespace, signer) + 1;
-
-                if (!_db.FileStorage.Exists(fileId))
-                {
-                    _db.FileStorage.Upload(
-                        fileId,
-                        signer.ToHex(),
-                        new MemoryStream());
-                }
-
-                LiteFileInfo file = _db.FileStorage.FindById(fileId);
-                using (var temp = new MemoryStream())
-                {
-                    file.CopyTo(temp);
-                    temp.Seek(0, SeekOrigin.Begin);
-                    byte[] prev = temp.ToArray();
-
-                    using (LiteFileStream stream = file.OpenWrite())
-                    {
-                        stream.Write(prev, 0, prev.Length);
-                        stream.Write(block.Hash.ToByteArray(), 0, hashSize);
-                        stream.Write(
-                            BitConverter.GetBytes(blockIndex), 0, sizeof(long));
-                        stream.Write(
-                            BitConverter.GetBytes(nextNonce), 0, sizeof(long));
-                    }
-                }
-            }
-        }
-
-        /// <inheritdoc/>
-        public void ForkTxNonce<T>(
-            string sourceNamespace,
-            string destinationNamespace,
-            Block<T> branchPoint,
-            IImmutableSet<Address> addressesToStrip)
-            where T : IAction, new()
-        {
-            long branchPointIndex = branchPoint.Index;
-            List<LiteFileInfo> files =
-                _db.FileStorage
-                    .Find($"{NonceIdPrefix}{sourceNamespace}")
-                    .ToList();
-
-            if (!files.Any() && addressesToStrip.Any())
-            {
-                throw new NamespaceNotFoundException(
-                    sourceNamespace,
-                    "The source namespace to be forked does not exist.");
-            }
-
-            foreach (LiteFileInfo srcFile in files)
-            {
-                string destId =
-                    $"{NonceIdPrefix}{destinationNamespace}/{srcFile.Filename}";
-                _db.FileStorage.Upload(
-                    destId,
-                    srcFile.Filename,
-                    new MemoryStream());
-
-                LiteFileInfo destFile = _db.FileStorage.FindById(destId);
-                using (LiteFileStream srcStream = srcFile.OpenRead())
-                using (LiteFileStream destStream = destFile.OpenWrite())
-                {
-                    while (srcStream.Position < srcStream.Length)
-                    {
-                        var hashBytes = new byte[HashDigest<SHA256>.Size];
-                        var indexBytes = new byte[sizeof(long)];
-                        var nonceBytes = new byte[sizeof(long)];
-
-                        srcStream.Read(hashBytes, 0, hashBytes.Length);
-                        srcStream.Read(indexBytes, 0, indexBytes.Length);
-                        srcStream.Read(nonceBytes, 0, nonceBytes.Length);
-
-                        long currentIndex =
-                            BitConverter.ToInt64(indexBytes, 0);
-
-                        if (currentIndex <= branchPointIndex)
-                        {
-                            destStream.Write(hashBytes, 0, hashBytes.Length);
-                            destStream.Write(indexBytes, 0, indexBytes.Length);
-                            destStream.Write(nonceBytes, 0, nonceBytes.Length);
-                        }
-                        else
-                        {
-                            break;
-                        }
-                    }
-                }
-
-                if (destFile.Length == 0)
-                {
-                    _db.FileStorage.Delete(destId);
-                }
-            }
+            long nextNonce = GetTxNonce(@namespace, signer) + delta;
+            var collectionId = $"{NonceIdPrefix}{@namespace}";
+            LiteCollection<BsonDocument> collection = _db.GetCollection<BsonDocument>(collectionId);
+            var docId = new BsonValue(signer.ToByteArray());
+            collection.Upsert(docId, new BsonDocument() { ["v"] = new BsonValue(nextNonce) });
         }
 
         /// <inheritdoc/>